add different secondary mapping count for seqs less than segLength
This helps us to map collections of reads with different sizes, where
some are below the desired segment length for the entire collection. We
can still use these sequences, but we might not want to map them many
places. This makes that configurable, so by default we at least align
them to their best mapping, but we can potentially multi-map too.