filter mappings with impossibly high identity estimates
Because we use a set of minimizers for the minimizer sketch, we are
vulnerable to an error where we map two sequences composed wholly or
mostly of the same repeat against each other even if their length is
dramatically different.

This filter removes alignments whose length difference would require an
identity lower than the -p[%], --map-pct-id=[%].

It also resolves a bug with block length calculation and identity
annotation for queries shorter than our segment length.