use edlib to get a more precise distance estimate
Before aligning each segment pair, we were checking if the distance was
acceptable using mash distance. This added some overhead to compute. It
is also problematically approximate, and added a new free parameter (the
kmer used) to the algorithm.

Instead, we can use edlib to get the edit distance. If this is
acceptable, we align as before. This is significantly faster and lets us
work with larger wflambda segments. On my test system, I find that
setting -W to 500 provides excellent runtime, so this has been set as
the default for wfmash.