take a little nibble to always patch with WFA
If WFA is run on short sequences, we can get traceback errors.  A simple
trick is to always give sequences of a given length to it for patching.
Doing this requires that we nibble forward a bit of match on the other
side of the region of the alignment matrix that we're patching. This
will mean we patch more in some cases, but in practice the cost seems to
be negligible.