keep the 99.5% least-frequent minimizers
We have to look at each kmer match, comparing the jaccard of minimizers
in query and target for each one. We should not be using transposons or
high-copy repeat kmers to do this. However, we can't discard all
repetitive kmers. The filter setting of 0.5% helps us avoid transposons
but retains sensitivity even to highly-repetitive sequences in segdups.