reduce to maintain the diagonal for pairs of differing lengths
This changes wflambdav2's reduction model to reduce relative to the
completion of the alignment in a given axis, irrespective of absolute
length. This ensures that we run down the central diagonal of the
alignment matrix even when the sequences are of very different lengths.

The specific math could still be improved, as distortion of the aligned
band does occur, with the band being larger at the beginning of the
alignment and smaller at the end.