tweaking recursive progressive patch
Chew back a tiny bit of the end of the alignment when finding bounds, which avoids getting confused by a single or handful of bp match at the beginning of the patch.

Only do recursive patching when we hit inversions (skip if we get just one forward alignment) because we only get overlapping alignments when we repeat the forward alignment multiple times.